refactor: move utility to utils
