+++ ietf/doc/views_doc.py
@@ -40,7 +40,6 @@ import os
 import re
 
 from pathlib import Path
-from typing import Optional
 
 from django.core.cache import caches
 from django.db.models import Max
@@ -66,7 +65,7 @@ from ietf.doc.utils import (augment_events_with_revision,
     add_events_message_info, get_unicode_document_content,
     augment_docs_and_person_with_person_info, irsg_needed_ballot_positions, add_action_holder_change_event,
     build_file_urls, update_documentauthors, fuzzy_find_documents,
-    bibxml_for_draft, DraftAliasGenerator)
+    bibxml_for_draft, get_doc_email_aliases)
 from ietf.doc.utils_bofreq import bofreq_editors, bofreq_responsible
 from ietf.group.models import Role, Group
 from ietf.group.utils import can_manage_all_groups_of_type, can_manage_materials, group_features_role_filter
@@ -1072,21 +1071,6 @@ def document_pdfized(request, name, rev=None, ext=None):
         return HttpResponse(pdf,content_type='application/pdf')
     else:
         raise Http404
-
-
-def get_doc_email_aliases(name: Optional[str] = None):
-    aliases = []
-    for (alias, alist) in DraftAliasGenerator(
-        Document.objects.filter(type_id="draft", name=name) if name else None
-    ):
-        # alias is draft-name.alias_type
-        doc_name, _dot, alias_type = alias.partition(".")
-        aliases.append({
-            "doc_name": doc_name,
-            "alias_type": f".{alias_type}" if alias_type else "",
-            "expansion": ", ".join(sorted(alist)),
-        })
-    return sorted(aliases, key=lambda a: (a["doc_name"]))
 
 
 def document_email(request,name):
